+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## $v
+
+- Bug fix. Regex patterns failed on bad SQL query.
+- Adding new option that allows a site owner to choose what their regex patterns will test; i.e., full URL, request URI, or just the path.
+
 ## v170730.42995
 
 - Initial release.